+++ cart/infrastructure/InMemoryBehaviour_test.go
@@ -49,7 +49,6 @@ func TestInMemoryBehaviour_CleanCart(t *testing.T) {
 				nil,
 				nil,
 				nil,
-				nil,
 			)
 			cart := &domaincart.Cart{
 				ID: "17",
@@ -177,7 +176,6 @@ func TestInMemoryBehaviour_CleanDelivery(t *testing.T) {
 				nil,
 				nil,
 				nil,
-				nil,
 			)
 			if err := cob.cartStorage.StoreCart(tt.args.cart); err != nil {
 				t.Fatalf("cart could not be initialized")
@@ -241,7 +239,6 @@ func TestInMemoryBehaviour_ApplyVoucher(t *testing.T) {
 				&InMemoryCartStorage{},
 				nil,
 				flamingo.NullLogger{},
-				nil,
 				nil,
 				nil,
 				nil,
@@ -339,7 +336,6 @@ func TestInMemoryBehaviour_RemoveVoucher(t *testing.T) {
 				func() *domaincart.Builder {
 					return &domaincart.Builder{}
 				},
-				nil,
 				&DefaultVoucherHandler{},
 				&DefaultGiftCardHandler{},
 				nil,
@@ -403,7 +399,6 @@ func TestInMemoryBehaviour_ApplyGiftCard(t *testing.T) {
 				&InMemoryCartStorage{},
 				nil,
 				flamingo.NullLogger{},
-				nil,
 				nil,
 				nil,
 				nil,
@@ -475,7 +470,6 @@ func TestInMemoryBehaviour_RemoveGiftCard(t *testing.T) {
 				nil,
 				nil,
 				nil,
-				nil,
 				&DefaultVoucherHandler{},
 				&DefaultGiftCardHandler{},
 				nil,
@@ -505,7 +499,6 @@ func TestInMemoryBehaviour_Complete(t *testing.T) {
 			nil,
 			nil,
 			nil,
-			nil,
 		)
 		cart := &domaincart.Cart{ID: "test-id"}
 		require.NoError(t, cob.storeCart(cart))
@@ -526,7 +519,6 @@ func TestInMemoryBehaviour_Restore(t *testing.T) {
 			&InMemoryCartStorage{},
 			nil,
 			flamingo.NullLogger{},
-			nil,
 			nil,
 			nil,
 			nil,